@@ -29,8 +29,8 @@ package com.st.BlueSTSDK;
 import android.bluetooth.BluetoothGattCharacteristic;
 import android.os.Handler;
 import android.os.HandlerThread;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.st.BlueSTSDK.Utils.BLENodeDefines;
 
