@@ -26,7 +26,7 @@
  ******************************************************************************/
 package com.st.BlueSTSDK.Features.Audio;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.st.BlueSTSDK.Feature;
 import com.st.BlueSTSDK.Features.Field;
