@@ -26,7 +26,7 @@
  ******************************************************************************/
 package com.st.BlueSTSDK.Features.Audio.Opus;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import com.st.BlueSTSDK.Features.Audio.AudioCodecManager;
