@@ -38,8 +38,8 @@
 package com.st.BlueSTSDK.Features;
 
 
-import android.support.annotation.IntDef;
-import android.util.Log;
+import androidx.annotation.IntDef;
+
 import android.util.SparseArray;
 
 import com.st.BlueSTSDK.Feature;
@@ -49,7 +49,6 @@ import com.st.BlueSTSDK.Utils.NumberConversion;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
