@@ -26,8 +26,8 @@
  ******************************************************************************/
 package com.st.BlueSTSDK.Utils;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.st.BlueSTSDK.BuildConfig;
 import com.st.BlueSTSDK.Node;
