@@ -38,25 +38,22 @@ import com.st.BlueSTSDK.Features.Audio.ADPCM.FeatureAudioADPCM;
 import com.st.BlueSTSDK.Features.Audio.ADPCM.FeatureAudioADPCMSync;
 import com.st.BlueSTSDK.Features.Audio.Opus.FeatureAudioOpus;
 import com.st.BlueSTSDK.Features.Audio.Opus.FeatureAudioOpusConf;
-import com.st.BlueSTSDK.Features.FeatureAudioSceneClassification;
+import com.st.BlueSTSDK.Features.FeatureAudioClassification;
 import com.st.BlueSTSDK.Features.FeatureBeamforming;
 import com.st.BlueSTSDK.Features.FeatureBattery;
 import com.st.BlueSTSDK.Features.FeatureCOSensor;
 import com.st.BlueSTSDK.Features.FeatureCarryPosition;
 import com.st.BlueSTSDK.Features.FeatureCompass;
-import com.st.BlueSTSDK.Features.FeatureDeskTypeDetection;
 import com.st.BlueSTSDK.Features.FeatureDirectionOfArrival;
 import com.st.BlueSTSDK.Features.FeatureEulerAngle;
 import com.st.BlueSTSDK.Features.FeatureEventCounter;
 import com.st.BlueSTSDK.Features.FeatureFitnessActivity;
 import com.st.BlueSTSDK.Features.FeatureFreeFall;
 import com.st.BlueSTSDK.Features.FeatureGyroscope;
-import com.st.BlueSTSDK.Features.FeatureGyroscopeNorm;
 import com.st.BlueSTSDK.Features.FeatureHumidity;
 import com.st.BlueSTSDK.Features.FeatureFFTAmplitude;
 import com.st.BlueSTSDK.Features.FeatureLuminosity;
 import com.st.BlueSTSDK.Features.FeatureMagnetometer;
-import com.st.BlueSTSDK.Features.FeatureMagnetometerNorm;
 import com.st.BlueSTSDK.Features.FeatureMemsGesture;
 import com.st.BlueSTSDK.Features.FeatureMemsSensorFusion;
 import com.st.BlueSTSDK.Features.FeatureMemsSensorFusionCompact;
@@ -296,7 +293,7 @@ public class BLENodeDefines {
         public static final SparseArray<Class<? extends Feature>> DEFAULT_MASK_TO_FEATURE =
                 new SparseArray<>();
 
-        public static final SparseArray<Class<? extends Feature>> SENSOT_TILE_BOX_MASK_TO_FEATURE =
+        public static final SparseArray<Class<? extends Feature>> SENSOR_TILE_BOX_MASK_TO_FEATURE =
                 new SparseArray<>();
 
         /**
@@ -350,46 +347,46 @@ public class BLENodeDefines {
             DEFAULT_MASK_TO_FEATURE.put(0x00000001, FeaturePedometer.class);
         }
 
-        private static void buildSensorTile101FeatureMask(){
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x80000000, FeatureFFTAmplitude.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x40000000, FeatureAudioADPCMSync.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x20000000, FeatureSwitch.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x10000000, FeatureMemsNorm.class);
+        private static void buildSensorTileBoxFeatureMask(){
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x80000000, FeatureFFTAmplitude.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x40000000, FeatureAudioADPCMSync.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x20000000, FeatureSwitch.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x10000000, FeatureMemsNorm.class);
 
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x08000000, FeatureAudioADPCM.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x04000000, FeatureMicLevel.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x02000000, FeatureProximity.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x01000000, FeatureLuminosity.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x08000000, FeatureAudioADPCM.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x04000000, FeatureMicLevel.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x02000000, FeatureAudioClassification.class);
+            //SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x01000000, FeatureLuminosity.class);
 
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00800000, FeatureAcceleration.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00400000, FeatureGyroscope.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00200000, FeatureMagnetometer.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00100000, FeaturePressure.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00800000, FeatureAcceleration.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00400000, FeatureGyroscope.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00200000, FeatureMagnetometer.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00100000, FeaturePressure.class);
 
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00080000, FeatureHumidity.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00040000, FeatureTemperature.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00020000, FeatureBattery.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00010000, FeatureTemperature.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00080000, FeatureHumidity.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00040000, FeatureTemperature.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00020000, FeatureBattery.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00010000, FeatureTemperature.class);
 
-            //SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00008000, FeatureGyroscopeNorm.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00004000, FeatureEulerAngle.class);
+            //SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00008000, FeatureGyroscopeNorm.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00004000, FeatureEulerAngle.class);
             //DEFAULT_MASK_TO_FEATURE.put(0x00002000, RFU); // stm32 ota reboot
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00001000, FeatureSDLogging.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00001000, FeatureSDLogging.class);
 
-            //SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000800, FeatureMagnetometerNorm.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000400, FeatureAccelerationEvent.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000200, FeatureEventCounter.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000100, FeatureMemsSensorFusionCompact.class);
+            //SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000800, FeatureMagnetometerNorm.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000400, FeatureAccelerationEvent.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000200, FeatureEventCounter.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000100, FeatureMemsSensorFusionCompact.class);
 
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000080, FeatureMemsSensorFusion.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000020, FeatureMotionIntensity.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000040, FeatureCompass.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000010, FeatureActivity.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000080, FeatureMemsSensorFusion.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000020, FeatureMotionIntensity.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000040, FeatureCompass.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000010, FeatureActivity.class);
 
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000008, FeatureCarryPosition.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000004, FeatureProximityGesture.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000002, FeatureMemsGesture.class);
-            SENSOT_TILE_BOX_MASK_TO_FEATURE.put(0x00000001, FeaturePedometer.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000008, FeatureCarryPosition.class);
+            //SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000004, FeatureProximityGesture.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000002, FeatureMemsGesture.class);
+            SENSOR_TILE_BOX_MASK_TO_FEATURE.put(0x00000001, FeaturePedometer.class);
         }
 
         @SafeVarargs
@@ -400,7 +397,7 @@ public class BLENodeDefines {
         private static void buildExtendedFeatureMask(){
             EXTENDED_FEATURE_MAP.put(buildExtendedFeatureCharacteristics(0x01), asList(FeatureAudioOpus.class));
             EXTENDED_FEATURE_MAP.put(buildExtendedFeatureCharacteristics(0x02), asList(FeatureAudioOpusConf.class));
-            EXTENDED_FEATURE_MAP.put(buildExtendedFeatureCharacteristics(0x03), asList(FeatureAudioSceneClassification.class));
+            EXTENDED_FEATURE_MAP.put(buildExtendedFeatureCharacteristics(0x03), asList(FeatureAudioClassification.class));
             EXTENDED_FEATURE_MAP.put(buildExtendedFeatureCharacteristics(0x04), asList(FeatureAILogging.class));
             EXTENDED_FEATURE_MAP.put(buildExtendedFeatureCharacteristics(0x05), asList(FeatureFFTAmplitude.class));
             EXTENDED_FEATURE_MAP.put(buildExtendedFeatureCharacteristics(0x06), asList(FeatureMotorTimeParameter.class));
@@ -416,7 +413,7 @@ public class BLENodeDefines {
             buildDefaultBaseFeatureMask();
             buildExtendedFeatureMask();
 
-            buildSensorTile101FeatureMask();
+            buildSensorTileBoxFeatureMask();
 
             Nucleo_Remote_Features.put(0x20000000, RemoteFeatureSwitch.class);
             Nucleo_Remote_Features.put(0x00100000, RemoteFeaturePressure.class);
